feat: Add Detection Sampling Module and ONNX Model Exporter
- Implemented DetectionCounter class for counting detections using YOLOv5n ONNX model.
- Added configuration types and constants for dataset curation.
- Created ONNX Model Exporter package for exporting PyTorch models to ONNX format.
- Included modules for exporting YOLOv5n and MobileNetV2 with verification and checksum utilities.
- Organized code structure for maintainability and clarity.
